backend/requirements.txt
fastapi>=0.121.0
uvicorn>=0.38.0
httpx>=0.28.0
python-dotenv>=1.2.0
pydantic>=2.12.0
pypdf>=5.0.0
